feat: route-handler enrich.test.ts mocks @ai-sdk/anthropic (closes #344)
